move factory to AWSCredentials
