--- core/src/main/java/com/amplifyframework/auth/AWSCredentialsFactory.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-package com.amplifyframework.auth
-
-import aws.smithy.kotlin.runtime.time.Instant
-
-/**
- * Factory to create [AWSCredentials]
- */
-object AWSCredentialsFactory {
-
-    fun createAWSCredentials(
-        accessKeyId: String?,
-        secretAccessKey: String?,
-        sessionToken: String?,
-        expiration: Long?,
-    ): AWSCredentials? {
-        return when {
-            accessKeyId == null || secretAccessKey == null -> {
-                null
-            }
-            sessionToken != null && expiration != null -> {
-                AWSTemporaryCredentials(
-                    accessKeyId,
-                    secretAccessKey,
-                    sessionToken,
-                    Instant.fromEpochSeconds(expiration)
-                )
-            }
-            else -> {
-                AWSCredentials(accessKeyId, secretAccessKey)
-            }
-        }
-    }
-}
